Integrated Smart Parking to the City scene.

+++ Game/Assets/Scripts/Parking/ParkingSpot.cs
@@ -3,6 +3,6 @@
 	public class ParkingSpot
 	{
 		public bool Occupied;
-		public int TimeElapsed;
+		public float TimeElapsed;
 	}
 }
